wasserstoff/AiInternTask/requirements.txt
streamlit
transformers
torch
Pillow
langchain
qdrant-client
python-docx
pandas
openpyxl
requests
